+++ LocadoraVeiculos/src/Modelo/Cliente.java
@@ -10,13 +10,13 @@ package Modelo;
  */
 public class Cliente {
     String nome;
-    int cpf;
+    long cpf;
     String endereco;
-    int telefone;
+    long telefone;
     String email;
 
     // Construtor
-    public Cliente(String nome, int cpf, String endereco, int telefone, String email) {
+    public Cliente(String nome, long cpf, String endereco, long telefone, String email) {
         this.nome = nome;
         this.cpf = cpf;
         this.endereco = endereco;
@@ -36,11 +36,11 @@ public class Cliente {
         this.nome = nome;
     }
     
-    public int getCpf() {
+    public long getCpf() {
         return cpf;
     }
 
-    public void setCpf(int cpf) {
+    public void setCpf(long cpf) {
         this.cpf = cpf;
     }
 
@@ -52,11 +52,11 @@ public class Cliente {
         this.endereco = endereco;
     }
 
-    public int getTelefone() {
+    public long getTelefone() {
         return telefone;
     }
 
-    public void setTelefone(int telefone) {
+    public void setTelefone(long telefone) {
         this.telefone = telefone;
     }
